refactor: component lifecycle operations
