--- internal/controller/components.go
@@ -1,822 +1,0 @@
-/*
-Copyright 2025.
-
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-
-    http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
-*/
-
-package controller
-
-import (
-	"context"
-	"fmt"
-	"maps"
-	"time"
-
-	"github.com/go-logr/logr"
-	appsv1 "k8s.io/api/apps/v1"
-	v1 "k8s.io/api/core/v1"
-	apierrors "k8s.io/apimachinery/pkg/api/errors"
-	"k8s.io/apimachinery/pkg/api/resource"
-	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	"k8s.io/apimachinery/pkg/types"
-	ctrl "sigs.k8s.io/controller-runtime"
-
-	etlv1alpha1 "github.com/glassflow/glassflow-etl-k8s-operator/api/v1alpha1"
-	"github.com/glassflow/glassflow-etl-k8s-operator/internal/constants"
-	"github.com/glassflow/glassflow-etl-k8s-operator/internal/errs"
-	"github.com/glassflow/glassflow-etl-k8s-operator/internal/utils"
-)
-
-// stopPipelineComponents stops all pipeline components in the correct order with pending message checks
-// nolint:gocyclo
-func (r *PipelineReconciler) stopPipelineComponents(ctx context.Context, log logr.Logger, p *etlv1alpha1.Pipeline) (ctrl.Result, error) {
-	namespace := r.getTargetNamespace(*p)
-
-	// Step 1: Stop Ingestor StatefulSets and headless Services
-	for i := range p.Spec.Ingestor.Streams {
-		resourceName := r.getResourceName(*p, fmt.Sprintf("%s-%d", constants.IngestorComponent, i))
-		deleted, err := r.isStatefulSetAbsent(ctx, namespace, resourceName)
-		if err != nil {
-			return ctrl.Result{}, fmt.Errorf("check ingestor statefulset %s: %w", resourceName, err)
-		}
-		if !deleted {
-			timedOut, _ := r.checkOperationTimeout(log, p)
-			if timedOut {
-				return r.handleOperationTimeout(ctx, log, p, constants.OperationStop)
-			}
-			log.Info("deleting ingestor statefulset", "statefulset", resourceName, "namespace", namespace)
-			var sts appsv1.StatefulSet
-			err := r.Get(ctx, types.NamespacedName{Namespace: namespace, Name: resourceName}, &sts)
-			if err != nil {
-				if !apierrors.IsNotFound(err) {
-					return ctrl.Result{}, fmt.Errorf("get ingestor statefulset %s: %w", resourceName, err)
-				}
-			} else {
-				err = r.deleteStatefulSet(ctx, &sts)
-				if err != nil {
-					return ctrl.Result{}, fmt.Errorf("delete ingestor statefulset %s: %w", resourceName, err)
-				}
-			}
-			return ctrl.Result{Requeue: true, RequeueAfter: time.Second}, nil
-		}
-		svcAbsent, err := r.isServiceAbsent(ctx, namespace, resourceName)
-		if err != nil {
-			return ctrl.Result{}, fmt.Errorf("check ingestor service %s: %w", resourceName, err)
-		}
-		if !svcAbsent {
-			timedOut, _ := r.checkOperationTimeout(log, p)
-			if timedOut {
-				return r.handleOperationTimeout(ctx, log, p, constants.OperationStop)
-			}
-			log.Info("deleting ingestor headless service", "service", resourceName, "namespace", namespace)
-			var svc v1.Service
-			err := r.Get(ctx, types.NamespacedName{Namespace: namespace, Name: resourceName}, &svc)
-			if err != nil {
-				if !apierrors.IsNotFound(err) {
-					return ctrl.Result{}, fmt.Errorf("get ingestor service %s: %w", resourceName, err)
-				}
-			} else {
-				err = r.deleteService(ctx, &svc)
-				if err != nil {
-					return ctrl.Result{}, fmt.Errorf("delete ingestor service %s: %w", resourceName, err)
-				}
-			}
-			return ctrl.Result{Requeue: true, RequeueAfter: time.Second}, nil
-		}
-	}
-
-	// Step 2: Stop Dedup StatefulSets
-	for i, stream := range p.Spec.Ingestor.Streams {
-		if stream.Deduplication == nil || !stream.Deduplication.Enabled {
-			continue
-		}
-
-		dedupName := r.getResourceName(*p, fmt.Sprintf("dedup-%d", i))
-
-		// Check for timeout
-		timedOut, _ := r.checkOperationTimeout(log, p)
-		if timedOut {
-			return r.handleOperationTimeout(ctx, log, p, constants.OperationStop)
-		}
-
-		// Check for pending messages first
-		err := r.checkDedupPendingMessages(ctx, *p, i)
-		if err != nil {
-			if errs.IsConsumerPendingMessagesError(err) {
-				log.Info("dedup has pending messages, requeuing", "pipeline_id", p.Spec.ID, "stream_index", i, "error", err.Error())
-				return ctrl.Result{Requeue: true, RequeueAfter: 10 * time.Second}, nil
-			}
-			return ctrl.Result{}, fmt.Errorf("check dedup pending messages for stream %d: %w", i, err)
-		}
-
-		deleted, err := r.isStatefulSetAbsent(ctx, namespace, dedupName)
-		if err != nil {
-			return ctrl.Result{}, fmt.Errorf("check dedup statefulset %s: %w", dedupName, err)
-		}
-		if !deleted {
-			log.Info("deleting dedup statefulset", "statefulset", dedupName, "namespace", namespace)
-			var sts appsv1.StatefulSet
-			err := r.Get(ctx, types.NamespacedName{Namespace: namespace, Name: dedupName}, &sts)
-			if err != nil {
-				if !apierrors.IsNotFound(err) {
-					return ctrl.Result{}, fmt.Errorf("get dedup statefulset %s: %w", dedupName, err)
-				}
-			} else {
-				err = r.deleteStatefulSet(ctx, &sts)
-				if err != nil {
-					return ctrl.Result{}, fmt.Errorf("delete dedup statefulset %s: %w", dedupName, err)
-				}
-			}
-			return ctrl.Result{Requeue: true, RequeueAfter: time.Second}, nil
-		} else {
-			log.Info("dedup statefulset is already deleted", "statefulset", dedupName, "namespace", namespace)
-		}
-	}
-
-	// Step 3: Check join and stop Join deployment (if enabled)
-	if p.Spec.Join.Enabled {
-		// Check for timeout before checking pending messages
-		timedOut, _ := r.checkOperationTimeout(log, p)
-		if timedOut {
-			return r.handleOperationTimeout(ctx, log, p, constants.OperationStop)
-		}
-
-		// Check for pending messages first
-		err := r.checkJoinPendingMessages(ctx, *p)
-		if err != nil {
-			if errs.IsConsumerPendingMessagesError(err) {
-				log.Info("join has pending messages, requeuing operation",
-					"pipeline_id", p.Spec.ID,
-					"error", err.Error())
-				return ctrl.Result{Requeue: true, RequeueAfter: 10 * time.Second}, nil
-			}
-			return ctrl.Result{}, fmt.Errorf("check join pending messages: %w", err)
-		}
-
-		deleted, err := r.isDeploymentAbsent(ctx, namespace, r.getResourceName(*p, constants.JoinComponent))
-		if err != nil {
-			return ctrl.Result{}, fmt.Errorf("check join deployment: %w", err)
-		}
-		if !deleted {
-			// Check for timeout before requeuing
-			timedOut, _ := r.checkOperationTimeout(log, p)
-			if timedOut {
-				return r.handleOperationTimeout(ctx, log, p, constants.OperationStop)
-			}
-
-			log.Info("deleting join deployment", "namespace", namespace)
-			var deployment appsv1.Deployment
-			err := r.Get(ctx, types.NamespacedName{Namespace: namespace, Name: r.getResourceName(*p, constants.JoinComponent)}, &deployment)
-			if err != nil {
-				if !apierrors.IsNotFound(err) {
-					return ctrl.Result{}, fmt.Errorf("get join deployment: %w", err)
-				}
-			} else {
-				err = r.deleteDeployment(ctx, &deployment)
-				if err != nil {
-					return ctrl.Result{}, fmt.Errorf("delete join deployment: %w", err)
-				}
-			}
-			// Requeue to wait for deployment to be fully deleted
-			return ctrl.Result{Requeue: true, RequeueAfter: time.Second}, nil
-		} else {
-			log.Info("join deployment is already deleted", "namespace", namespace)
-		}
-	}
-
-	// Step 4: Check sink and stop Sink StatefulSet and headless Service
-	timedOut, _ := r.checkOperationTimeout(log, p)
-	if timedOut {
-		return r.handleOperationTimeout(ctx, log, p, constants.OperationStop)
-	}
-	err := r.checkSinkPendingMessages(ctx, *p)
-	if err != nil {
-		if errs.IsConsumerPendingMessagesError(err) {
-			log.Info("sink has pending messages, requeuing operation",
-				"pipeline_id", p.Spec.ID,
-				"error", err.Error())
-			return ctrl.Result{Requeue: true, RequeueAfter: 10 * time.Second}, nil
-		}
-		return ctrl.Result{}, fmt.Errorf("check sink pending messages: %w", err)
-	}
-
-	sinkName := r.getResourceName(*p, constants.SinkComponent)
-	deleted, err := r.isStatefulSetAbsent(ctx, namespace, sinkName)
-	if err != nil {
-		return ctrl.Result{}, fmt.Errorf("check sink statefulset: %w", err)
-	}
-	if !deleted {
-		timedOut, _ := r.checkOperationTimeout(log, p)
-		if timedOut {
-			return r.handleOperationTimeout(ctx, log, p, constants.OperationStop)
-		}
-		log.Info("deleting sink statefulset", "namespace", namespace)
-		var sts appsv1.StatefulSet
-		err := r.Get(ctx, types.NamespacedName{Namespace: namespace, Name: sinkName}, &sts)
-		if err != nil {
-			if !apierrors.IsNotFound(err) {
-				return ctrl.Result{}, fmt.Errorf("get sink statefulset: %w", err)
-			}
-		} else {
-			err = r.deleteStatefulSet(ctx, &sts)
-			if err != nil {
-				return ctrl.Result{}, fmt.Errorf("delete sink statefulset: %w", err)
-			}
-		}
-		return ctrl.Result{Requeue: true, RequeueAfter: time.Second}, nil
-	}
-	svcAbsent, err := r.isServiceAbsent(ctx, namespace, sinkName)
-	if err != nil {
-		return ctrl.Result{}, fmt.Errorf("check sink service: %w", err)
-	}
-	if !svcAbsent {
-		timedOut, _ := r.checkOperationTimeout(log, p)
-		if timedOut {
-			return r.handleOperationTimeout(ctx, log, p, constants.OperationStop)
-		}
-		log.Info("deleting sink headless service", "namespace", namespace)
-		var svc v1.Service
-		err := r.Get(ctx, types.NamespacedName{Namespace: namespace, Name: sinkName}, &svc)
-		if err != nil {
-			if !apierrors.IsNotFound(err) {
-				return ctrl.Result{}, fmt.Errorf("get sink service: %w", err)
-			}
-		} else {
-			err = r.deleteService(ctx, &svc)
-			if err != nil {
-				return ctrl.Result{}, fmt.Errorf("delete sink service: %w", err)
-			}
-		}
-		return ctrl.Result{Requeue: true, RequeueAfter: time.Second}, nil
-	}
-
-	// All components (ingestor/sink StatefulSets and Services, Join deployment) are deleted
-	return ctrl.Result{}, nil
-}
-
-// terminatePipelineComponents terminates all pipeline components immediately
-func (r *PipelineReconciler) terminatePipelineComponents(ctx context.Context, log logr.Logger, p etlv1alpha1.Pipeline) (ctrl.Result, error) {
-	namespace := r.getTargetNamespace(p)
-
-	result, requeue, err := r.terminateIngestors(ctx, log, p, namespace)
-	if err != nil || requeue {
-		return result, err
-	}
-
-	result, requeue, err = r.terminateDedups(ctx, log, p, namespace)
-	if err != nil || requeue {
-		return result, err
-	}
-
-	result, requeue, err = r.terminateJoin(ctx, log, p, namespace)
-	if err != nil || requeue {
-		return result, err
-	}
-
-	result, requeue, err = r.terminateSink(ctx, log, p, namespace)
-	if err != nil || requeue {
-		return result, err
-	}
-
-	// All components are deleted
-	return ctrl.Result{}, nil
-}
-
-func (r *PipelineReconciler) terminateIngestors(ctx context.Context, log logr.Logger, p etlv1alpha1.Pipeline, namespace string) (ctrl.Result, bool, error) {
-	for i := range p.Spec.Ingestor.Streams {
-		resourceName := r.getResourceName(p, fmt.Sprintf("%s-%d", constants.IngestorComponent, i))
-		deleted, err := r.isStatefulSetAbsent(ctx, namespace, resourceName)
-		if err != nil {
-			return ctrl.Result{}, false, fmt.Errorf("check ingestor statefulset %s: %w", resourceName, err)
-		}
-		if !deleted {
-			log.Info("deleting ingestor statefulset", "statefulset", resourceName, "namespace", namespace)
-			var sts appsv1.StatefulSet
-			err := r.Get(ctx, types.NamespacedName{Namespace: namespace, Name: resourceName}, &sts)
-			if err != nil {
-				if !apierrors.IsNotFound(err) {
-					return ctrl.Result{}, false, fmt.Errorf("get ingestor statefulset %s: %w", resourceName, err)
-				}
-			} else {
-				if err = r.deleteStatefulSet(ctx, &sts); err != nil {
-					return ctrl.Result{}, false, fmt.Errorf("delete ingestor statefulset %s: %w", resourceName, err)
-				}
-			}
-			return ctrl.Result{Requeue: true, RequeueAfter: time.Second}, true, nil
-		}
-
-		svcAbsent, err := r.isServiceAbsent(ctx, namespace, resourceName)
-		if err != nil {
-			return ctrl.Result{}, false, fmt.Errorf("check ingestor service %s: %w", resourceName, err)
-		}
-		if !svcAbsent {
-			log.Info("deleting ingestor headless service", "service", resourceName, "namespace", namespace)
-			var svc v1.Service
-			err := r.Get(ctx, types.NamespacedName{Namespace: namespace, Name: resourceName}, &svc)
-			if err != nil {
-				if !apierrors.IsNotFound(err) {
-					return ctrl.Result{}, false, fmt.Errorf("get ingestor service %s: %w", resourceName, err)
-				}
-			} else {
-				if err = r.deleteService(ctx, &svc); err != nil {
-					return ctrl.Result{}, false, fmt.Errorf("delete ingestor service %s: %w", resourceName, err)
-				}
-			}
-			return ctrl.Result{Requeue: true, RequeueAfter: time.Second}, true, nil
-		}
-	}
-
-	return ctrl.Result{}, false, nil
-}
-
-func (r *PipelineReconciler) terminateDedups(ctx context.Context, log logr.Logger, p etlv1alpha1.Pipeline, namespace string) (ctrl.Result, bool, error) {
-	for i, stream := range p.Spec.Ingestor.Streams {
-		if stream.Deduplication == nil || !stream.Deduplication.Enabled {
-			continue
-		}
-
-		dedupName := r.getResourceName(p, fmt.Sprintf("dedup-%d", i))
-		deleted, err := r.isStatefulSetAbsent(ctx, namespace, dedupName)
-		if err != nil {
-			return ctrl.Result{}, false, fmt.Errorf("check dedup statefulset %s: %w", dedupName, err)
-		}
-		if !deleted {
-			log.Info("deleting dedup statefulset", "statefulset", dedupName, "namespace", namespace)
-			var sts appsv1.StatefulSet
-			err := r.Get(ctx, types.NamespacedName{Namespace: namespace, Name: dedupName}, &sts)
-			if err != nil {
-				if !apierrors.IsNotFound(err) {
-					return ctrl.Result{}, false, fmt.Errorf("get dedup statefulset %s: %w", dedupName, err)
-				}
-			} else {
-				if err = r.deleteStatefulSet(ctx, &sts); err != nil {
-					return ctrl.Result{}, false, fmt.Errorf("delete dedup statefulset %s: %w", dedupName, err)
-				}
-			}
-			return ctrl.Result{Requeue: true, RequeueAfter: time.Second}, true, nil
-		}
-
-		log.Info("dedup statefulset is already deleted", "statefulset", dedupName, "namespace", namespace)
-	}
-
-	return ctrl.Result{}, false, nil
-}
-
-func (r *PipelineReconciler) terminateJoin(ctx context.Context, log logr.Logger, p etlv1alpha1.Pipeline, namespace string) (ctrl.Result, bool, error) {
-	if !p.Spec.Join.Enabled {
-		return ctrl.Result{}, false, nil
-	}
-
-	joinName := r.getResourceName(p, constants.JoinComponent)
-	deleted, err := r.isDeploymentAbsent(ctx, namespace, joinName)
-	if err != nil {
-		return ctrl.Result{}, false, fmt.Errorf("check join deployment: %w", err)
-	}
-	if deleted {
-		log.Info("join deployment is already deleted", "namespace", namespace)
-		return ctrl.Result{}, false, nil
-	}
-
-	log.Info("deleting join deployment", "namespace", namespace)
-	var deployment appsv1.Deployment
-	err = r.Get(ctx, types.NamespacedName{Namespace: namespace, Name: joinName}, &deployment)
-	if err != nil {
-		if !apierrors.IsNotFound(err) {
-			return ctrl.Result{}, false, fmt.Errorf("get join deployment: %w", err)
-		}
-	} else {
-		if err = r.deleteDeployment(ctx, &deployment); err != nil {
-			return ctrl.Result{}, false, fmt.Errorf("delete join deployment: %w", err)
-		}
-	}
-
-	return ctrl.Result{Requeue: true, RequeueAfter: time.Second}, true, nil
-}
-
-func (r *PipelineReconciler) terminateSink(ctx context.Context, log logr.Logger, p etlv1alpha1.Pipeline, namespace string) (ctrl.Result, bool, error) {
-	sinkName := r.getResourceName(p, constants.SinkComponent)
-	deleted, err := r.isStatefulSetAbsent(ctx, namespace, sinkName)
-	if err != nil {
-		return ctrl.Result{}, false, fmt.Errorf("check sink statefulset: %w", err)
-	}
-	if !deleted {
-		log.Info("deleting sink statefulset", "namespace", namespace)
-		var sts appsv1.StatefulSet
-		err := r.Get(ctx, types.NamespacedName{Namespace: namespace, Name: sinkName}, &sts)
-		if err != nil {
-			if !apierrors.IsNotFound(err) {
-				return ctrl.Result{}, false, fmt.Errorf("get sink statefulset: %w", err)
-			}
-		} else {
-			if err = r.deleteStatefulSet(ctx, &sts); err != nil {
-				return ctrl.Result{}, false, fmt.Errorf("delete sink statefulset: %w", err)
-			}
-		}
-		return ctrl.Result{Requeue: true, RequeueAfter: time.Second}, true, nil
-	}
-
-	svcAbsent, err := r.isServiceAbsent(ctx, namespace, sinkName)
-	if err != nil {
-		return ctrl.Result{}, false, fmt.Errorf("check sink service: %w", err)
-	}
-	if svcAbsent {
-		return ctrl.Result{}, false, nil
-	}
-
-	log.Info("deleting sink headless service", "namespace", namespace)
-	var svc v1.Service
-	err = r.Get(ctx, types.NamespacedName{Namespace: namespace, Name: sinkName}, &svc)
-	if err != nil {
-		if !apierrors.IsNotFound(err) {
-			return ctrl.Result{}, false, fmt.Errorf("get sink service: %w", err)
-		}
-	} else {
-		if err = r.deleteService(ctx, &svc); err != nil {
-			return ctrl.Result{}, false, fmt.Errorf("delete sink service: %w", err)
-		}
-	}
-
-	return ctrl.Result{Requeue: true, RequeueAfter: time.Second}, true, nil
-}
-
-// createIngestors creates ingestor StatefulSets (and headless Services) for the pipeline.
-func (r *PipelineReconciler) createIngestors(ctx context.Context, _ logr.Logger, ns v1.Namespace, labels map[string]string, secret v1.Secret, p etlv1alpha1.Pipeline) error {
-	ing := p.Spec.Ingestor
-	namespace := ns.GetName()
-
-	for i, t := range ing.Streams {
-		resourceRef := r.getResourceName(p, fmt.Sprintf("%s-%d", constants.IngestorComponent, i))
-
-		ingestorLabels := r.getKafkaIngestorLabels(t.TopicName)
-		maps.Copy(ingestorLabels, labels)
-
-		err := r.createHeadlessService(ctx, namespace, resourceRef, ingestorLabels)
-		if err != nil {
-			return fmt.Errorf("create ingestor headless service %s: %w", resourceRef, err)
-		}
-
-		containerBuilder := newComponentContainerBuilder().
-			withName(resourceRef).
-			withImage(r.IngestorImage).
-			withImagePullPolicy(r.IngestorPullPolicy).
-			withVolumeMount(v1.VolumeMount{
-				Name:      "config",
-				ReadOnly:  true,
-				MountPath: "/config",
-			}).
-			withEnv(append(append(append(append([]v1.EnvVar{
-				{Name: "GLASSFLOW_NATS_SERVER", Value: r.ComponentNATSAddr},
-				{Name: "GLASSFLOW_PIPELINE_CONFIG", Value: "/config/pipeline.json"},
-				{Name: "GLASSFLOW_INGESTOR_TOPIC", Value: t.TopicName},
-				{Name: "NATS_SUBJECT_PREFIX", Value: getIngestorSubjectPrefix(p.Spec.ID, t.TopicName)},
-				{Name: "GLASSFLOW_LOG_LEVEL", Value: r.IngestorLogLevel},
-				{Name: "GLASSFLOW_NATS_MAX_STREAM_AGE", Value: r.NATSMaxStreamAge},
-				{Name: "GLASSFLOW_NATS_MAX_STREAM_BYTES", Value: utils.ConvertBytesToString(r.NATSMaxStreamBytes)},
-
-				{Name: "GLASSFLOW_OTEL_LOGS_ENABLED", Value: r.ObservabilityLogsEnabled},
-				{Name: "GLASSFLOW_OTEL_METRICS_ENABLED", Value: r.ObservabilityMetricsEnabled},
-				{Name: "OTEL_EXPORTER_OTLP_ENDPOINT", Value: r.ObservabilityOTelEndpoint},
-				{Name: "GLASSFLOW_OTEL_SERVICE_NAME", Value: constants.IngestorComponent},
-				{Name: "GLASSFLOW_OTEL_SERVICE_VERSION", Value: r.IngestorImageTag},
-				{Name: "GLASSFLOW_OTEL_SERVICE_NAMESPACE", Value: r.getTargetNamespace(p)},
-				{Name: "GLASSFLOW_OTEL_PIPELINE_ID", Value: p.Spec.ID},
-				{Name: "GLASSFLOW_OTEL_SERVICE_INSTANCE_ID", ValueFrom: &v1.EnvVarSource{
-					FieldRef: &v1.ObjectFieldSelector{
-						FieldPath: "metadata.name",
-					},
-				}},
-			}, ingestorNATSSubjectCountEnvVars(t)...), r.getStatefulSetPodIdentityEnvVars()...), r.getComponentDatabaseEnvVars()...), r.getUsageStatsEnvVars()...)).
-			withResources(r.IngestorCPURequest, r.IngestorCPULimit, r.IngestorMemoryRequest, r.IngestorMemoryLimit)
-		if mount, ok := r.getComponentEncryptionVolumeMount(); ok {
-			containerBuilder = containerBuilder.withVolumeMount(mount)
-		}
-		container := containerBuilder.build()
-
-		stsBuilder := newComponentStatefulSetBuilder().
-			withNamespace(ns).
-			withResourceName(resourceRef).
-			withServiceName(resourceRef).
-			withLabels(ingestorLabels).
-			withVolume(v1.Volume{
-				Name: "config",
-				VolumeSource: v1.VolumeSource{
-					Secret: &v1.SecretVolumeSource{
-						SecretName:  secret.Name,
-						DefaultMode: ptrInt32(0o600),
-					},
-				},
-			}).
-			withReplicas(t.Replicas).
-			withContainer(*container).
-			withAffinity(r.IngestorAffinity)
-		if vol, ok := r.getComponentEncryptionVolume(); ok {
-			stsBuilder = stsBuilder.withVolume(vol)
-		}
-		statefulSet := stsBuilder.build()
-
-		err = r.createStatefulSet(ctx, statefulSet)
-		if err != nil {
-			return fmt.Errorf("create ingestor statefulset %s: %w", resourceRef, err)
-		}
-	}
-
-	return nil
-}
-
-// createJoin creates a join deployment for the pipeline
-func (r *PipelineReconciler) createJoin(ctx context.Context, ns v1.Namespace, labels map[string]string, secret v1.Secret, p etlv1alpha1.Pipeline) error {
-	resourceRef := r.getResourceName(p, constants.JoinComponent)
-
-	joinLabels := r.getJoinLabels()
-
-	maps.Copy(joinLabels, labels)
-
-	joinContainerBuilder := newComponentContainerBuilder().
-		withName(resourceRef).
-		withImage(r.JoinImage).
-		withImagePullPolicy(r.JoinPullPolicy).
-		withVolumeMount(v1.VolumeMount{
-			Name:      "config",
-			ReadOnly:  true,
-			MountPath: "/config",
-		}).
-		withEnv(append(append([]v1.EnvVar{
-			{Name: "GLASSFLOW_NATS_SERVER", Value: r.ComponentNATSAddr},
-			{Name: "GLASSFLOW_PIPELINE_CONFIG", Value: "/config/pipeline.json"},
-			{Name: "GLASSFLOW_LOG_LEVEL", Value: r.JoinLogLevel},
-			{Name: "GLASSFLOW_NATS_MAX_STREAM_AGE", Value: r.NATSMaxStreamAge},
-			{Name: "GLASSFLOW_NATS_MAX_STREAM_BYTES", Value: utils.ConvertBytesToString(r.NATSMaxStreamBytes)},
-
-			{Name: "GLASSFLOW_OTEL_LOGS_ENABLED", Value: r.ObservabilityLogsEnabled},
-			{Name: "GLASSFLOW_OTEL_METRICS_ENABLED", Value: r.ObservabilityMetricsEnabled},
-			{Name: "OTEL_EXPORTER_OTLP_ENDPOINT", Value: r.ObservabilityOTelEndpoint},
-			{Name: "GLASSFLOW_OTEL_SERVICE_NAME", Value: constants.JoinComponent},
-			{Name: "GLASSFLOW_OTEL_SERVICE_VERSION", Value: r.JoinImageTag},
-			{Name: "GLASSFLOW_OTEL_SERVICE_NAMESPACE", Value: r.getTargetNamespace(p)},
-			{Name: "GLASSFLOW_OTEL_PIPELINE_ID", Value: p.Spec.ID},
-			{Name: "GLASSFLOW_OTEL_SERVICE_INSTANCE_ID", ValueFrom: &v1.EnvVarSource{
-				FieldRef: &v1.ObjectFieldSelector{
-					FieldPath: "metadata.name",
-				},
-			}},
-		}, r.getComponentDatabaseEnvVars()...), r.getUsageStatsEnvVars()...)).
-		withResources(r.JoinCPURequest, r.JoinCPULimit, r.JoinMemoryRequest, r.JoinMemoryLimit)
-	if mount, ok := r.getComponentEncryptionVolumeMount(); ok {
-		joinContainerBuilder = joinContainerBuilder.withVolumeMount(mount)
-	}
-	joinContainer := joinContainerBuilder.build()
-
-	joinDepBuilder := newComponentDeploymentBuilder().
-		withNamespace(ns).
-		withResourceName(resourceRef).
-		withLabels(joinLabels).
-		withVolume(v1.Volume{
-			Name: "config",
-			VolumeSource: v1.VolumeSource{
-				Secret: &v1.SecretVolumeSource{
-					SecretName:  secret.Name,
-					DefaultMode: ptrInt32(0o600),
-				},
-			},
-		}).
-		withContainer(*joinContainer).
-		withAffinity(r.JoinAffinity)
-	if vol, ok := r.getComponentEncryptionVolume(); ok {
-		joinDepBuilder = joinDepBuilder.withVolume(vol)
-	}
-	deployment := joinDepBuilder.build()
-
-	err := r.createDeployment(ctx, deployment)
-	if err != nil {
-		return fmt.Errorf("create join deployment: %w", err)
-	}
-
-	return nil
-}
-
-// createSink creates a sink StatefulSet (and headless Service) for the pipeline.
-func (r *PipelineReconciler) createSink(ctx context.Context, ns v1.Namespace, labels map[string]string, secret v1.Secret, p etlv1alpha1.Pipeline) error {
-	resourceRef := r.getResourceName(p, constants.SinkComponent)
-	namespace := ns.GetName()
-
-	sinkLabels := r.getSinkLabels()
-	maps.Copy(sinkLabels, labels)
-
-	err := r.createHeadlessService(ctx, namespace, resourceRef, sinkLabels)
-	if err != nil {
-		return fmt.Errorf("create sink headless service: %w", err)
-	}
-
-	sinkContainerBuilder := newComponentContainerBuilder().
-		withName(resourceRef).
-		withImage(r.SinkImage).
-		withImagePullPolicy(r.SinkPullPolicy).
-		withVolumeMount(v1.VolumeMount{
-			Name:      "config",
-			ReadOnly:  true,
-			MountPath: "/config",
-		}).
-		withEnv(append(append(append([]v1.EnvVar{
-			{Name: "GLASSFLOW_NATS_SERVER", Value: r.ComponentNATSAddr},
-			{Name: "GLASSFLOW_PIPELINE_CONFIG", Value: "/config/pipeline.json"},
-			{Name: "NATS_INPUT_STREAM_PREFIX", Value: getSinkInputStreamPrefix(p.Spec.ID)},
-			{Name: "GLASSFLOW_LOG_LEVEL", Value: r.SinkLogLevel},
-			{Name: "GLASSFLOW_NATS_MAX_STREAM_AGE", Value: r.NATSMaxStreamAge},
-			{Name: "GLASSFLOW_NATS_MAX_STREAM_BYTES", Value: utils.ConvertBytesToString(r.NATSMaxStreamBytes)},
-
-			{Name: "GLASSFLOW_OTEL_LOGS_ENABLED", Value: r.ObservabilityLogsEnabled},
-			{Name: "GLASSFLOW_OTEL_METRICS_ENABLED", Value: r.ObservabilityMetricsEnabled},
-			{Name: "OTEL_EXPORTER_OTLP_ENDPOINT", Value: r.ObservabilityOTelEndpoint},
-			{Name: "GLASSFLOW_OTEL_SERVICE_NAME", Value: constants.SinkComponent},
-			{Name: "GLASSFLOW_OTEL_SERVICE_VERSION", Value: r.SinkImageTag},
-			{Name: "GLASSFLOW_OTEL_SERVICE_NAMESPACE", Value: r.getTargetNamespace(p)},
-			{Name: "GLASSFLOW_OTEL_PIPELINE_ID", Value: p.Spec.ID},
-			{Name: "GLASSFLOW_OTEL_SERVICE_INSTANCE_ID", ValueFrom: &v1.EnvVarSource{
-				FieldRef: &v1.ObjectFieldSelector{
-					FieldPath: "metadata.name",
-				},
-			}},
-		}, r.getStatefulSetPodIdentityEnvVars()...), r.getComponentDatabaseEnvVars()...), r.getUsageStatsEnvVars()...)).
-		withResources(r.SinkCPURequest, r.SinkCPULimit, r.SinkMemoryRequest, r.SinkMemoryLimit)
-	if mount, ok := r.getComponentEncryptionVolumeMount(); ok {
-		sinkContainerBuilder = sinkContainerBuilder.withVolumeMount(mount)
-	}
-	sinkContainer := sinkContainerBuilder.build()
-
-	stsBuilder := newComponentStatefulSetBuilder().
-		withNamespace(ns).
-		withResourceName(resourceRef).
-		withServiceName(resourceRef).
-		withLabels(sinkLabels).
-		withVolume(v1.Volume{
-			Name: "config",
-			VolumeSource: v1.VolumeSource{
-				Secret: &v1.SecretVolumeSource{
-					SecretName:  secret.Name,
-					DefaultMode: ptrInt32(0o600),
-				},
-			},
-		}).
-		withContainer(*sinkContainer).
-		withReplicas(getSinkReplicaCount(p)).
-		withAffinity(r.SinkAffinity)
-	if vol, ok := r.getComponentEncryptionVolume(); ok {
-		stsBuilder = stsBuilder.withVolume(vol)
-	}
-	statefulSet := stsBuilder.build()
-
-	err = r.createStatefulSet(ctx, statefulSet)
-	if err != nil {
-		return fmt.Errorf("create sink statefulset: %w", err)
-	}
-
-	return nil
-}
-
-// createDedups creates dedup StatefulSets for the pipeline (one per ingestor stream with dedup enabled)
-func (r *PipelineReconciler) createDedups(ctx context.Context, _ logr.Logger, ns v1.Namespace, labels map[string]string, secret v1.Secret, p etlv1alpha1.Pipeline) error {
-	ing := p.Spec.Ingestor
-
-	for i, stream := range ing.Streams {
-		// Skip if dedup not enabled for this stream
-		if stream.Deduplication == nil || !stream.Deduplication.Enabled {
-			continue
-		}
-
-		resourceRef := r.getResourceName(p, fmt.Sprintf("dedup-%d", i))
-		serviceName := resourceRef
-
-		dedupLabels := r.getDedupLabels(stream.TopicName)
-		maps.Copy(dedupLabels, labels)
-
-		replicas := getDedupReplicaCount(stream)
-
-		// Determine storage size (use pipeline config, fallback to Helm default)
-		storageSize := r.DedupDefaultStorageSize
-		if stream.Deduplication.StorageSize != "" {
-			storageSize = stream.Deduplication.StorageSize
-		}
-
-		// Determine storage class (use pipeline config, fallback to Helm default)
-		storageClass := r.DedupDefaultStorageClass
-		if stream.Deduplication.StorageClass != "" {
-			storageClass = stream.Deduplication.StorageClass
-		}
-
-		dedupEnvBase := []v1.EnvVar{
-			{Name: "GLASSFLOW_NATS_SERVER", Value: r.ComponentNATSAddr},
-			{Name: "GLASSFLOW_PIPELINE_CONFIG", Value: "/config/pipeline.json"},
-			{Name: "GLASSFLOW_DEDUP_TOPIC", Value: stream.TopicName},
-			{Name: "GLASSFLOW_SOURCE_INDEX", Value: fmt.Sprintf("%d", i)},
-			{Name: "GLASSFLOW_BADGER_PATH", Value: "/data/badger"},
-			{Name: "GLASSFLOW_LOG_LEVEL", Value: r.DedupLogLevel},
-			{Name: "GLASSFLOW_NATS_MAX_STREAM_AGE", Value: r.NATSMaxStreamAge},
-			{Name: "GLASSFLOW_NATS_MAX_STREAM_BYTES", Value: utils.ConvertBytesToString(r.NATSMaxStreamBytes)},
-
-			{Name: "GLASSFLOW_OTEL_LOGS_ENABLED", Value: r.ObservabilityLogsEnabled},
-			{Name: "GLASSFLOW_OTEL_METRICS_ENABLED", Value: r.ObservabilityMetricsEnabled},
-			{Name: "OTEL_EXPORTER_OTLP_ENDPOINT", Value: r.ObservabilityOTelEndpoint},
-			{Name: "GLASSFLOW_OTEL_SERVICE_NAME", Value: constants.DedupComponent},
-			{Name: "GLASSFLOW_OTEL_SERVICE_VERSION", Value: r.DedupImageTag},
-			{Name: "GLASSFLOW_OTEL_SERVICE_NAMESPACE", Value: r.getTargetNamespace(p)},
-			{Name: "GLASSFLOW_OTEL_PIPELINE_ID", Value: p.Spec.ID},
-			{Name: "GLASSFLOW_OTEL_SERVICE_INSTANCE_ID", ValueFrom: &v1.EnvVarSource{
-				FieldRef: &v1.ObjectFieldSelector{
-					FieldPath: "metadata.name",
-				},
-			}},
-		}
-		if useDedupNStreamPath(p) {
-			dedupEnvBase = append(dedupEnvBase,
-				v1.EnvVar{Name: "NATS_INPUT_STREAM_PREFIX", Value: getDedupInputStreamPrefix(p.Spec.ID, stream.TopicName)},
-				v1.EnvVar{Name: "NATS_SUBJECT_PREFIX", Value: getDedupOutputSubjectPrefix(p.Spec.ID, stream.TopicName)},
-			)
-		} else {
-			dedupEnvBase = append(dedupEnvBase,
-				v1.EnvVar{Name: "GLASSFLOW_INPUT_STREAM", Value: stream.OutputStream},
-				v1.EnvVar{Name: "GLASSFLOW_OUTPUT_STREAM", Value: stream.Deduplication.OutputStream},
-			)
-		}
-		dedupContainerBuilder := newComponentContainerBuilder().
-			withName(resourceRef).
-			withImage(r.DedupImage).
-			withImagePullPolicy(r.DedupPullPolicy).
-			withVolumeMount(v1.VolumeMount{
-				Name:      "config",
-				ReadOnly:  true,
-				MountPath: "/config",
-			}).
-			withVolumeMount(v1.VolumeMount{
-				Name:      "data",
-				MountPath: "/data/badger",
-			}).
-			withEnv(append(append(append(dedupEnvBase, r.getStatefulSetPodIdentityEnvVars()...), r.getComponentDatabaseEnvVars()...), r.getUsageStatsEnvVars()...)).
-			withResources(r.DedupCPURequest, r.DedupCPULimit, r.DedupMemoryRequest, r.DedupMemoryLimit)
-		if mount, ok := r.getComponentEncryptionVolumeMount(); ok {
-			dedupContainerBuilder = dedupContainerBuilder.withVolumeMount(mount)
-		}
-		container := dedupContainerBuilder.build()
-
-		// Parse storage size
-		storageSizeQuantity, err := resource.ParseQuantity(storageSize)
-		if err != nil {
-			return fmt.Errorf("parse storage size for dedup-%d: %w", i, err)
-		}
-
-		// Create PVC template
-		pvcTemplate := v1.PersistentVolumeClaim{
-			ObjectMeta: metav1.ObjectMeta{Name: "data"},
-			Spec: v1.PersistentVolumeClaimSpec{
-				AccessModes: []v1.PersistentVolumeAccessMode{v1.ReadWriteOnce},
-				Resources: v1.VolumeResourceRequirements{
-					Requests: v1.ResourceList{v1.ResourceStorage: storageSizeQuantity},
-				},
-			},
-		}
-
-		if storageClass != "" {
-			pvcTemplate.Spec.StorageClassName = &storageClass
-		}
-
-		stsBuilder := newComponentStatefulSetBuilder().
-			withNamespace(ns).
-			withResourceName(resourceRef).
-			withServiceName(serviceName).
-			withLabels(dedupLabels).
-			withVolume(v1.Volume{
-				Name: "config",
-				VolumeSource: v1.VolumeSource{
-					Secret: &v1.SecretVolumeSource{
-						SecretName:  secret.Name,
-						DefaultMode: ptrInt32(0o600),
-					},
-				},
-			}).
-			withReplicas(replicas).
-			withContainer(*container).
-			withAffinity(r.DedupAffinity).
-			withVolumeClaimTemplate(pvcTemplate)
-		if vol, ok := r.getComponentEncryptionVolume(); ok {
-			stsBuilder = stsBuilder.withVolume(vol)
-		}
-		statefulSet := stsBuilder.build()
-
-		err = r.createStatefulSet(ctx, statefulSet)
-		if err != nil {
-			return fmt.Errorf("create dedup-%d statefulset: %w", i, err)
-		}
-
-	}
-
-	return nil
-}
